Update shared files to rra-c-util 9.0
This is eight years of updates to rra-c-util support files for Perl
modules, so there are too many changes to list separately.

Stop using the Readonly module in favor of simple my-scoped global
variables, which are good enough and don't have the overhead of
requiring another CPAN module or using the confusing syntax of
use constant.

Adjust coding style to satisfy Perl::Critic::Freenode and current
perltidy.