Simplify the build system
Stop using a Module::Build add-on to automatically set the version
of modules to match the release, since this confuses the CPAN
indexer.  Instead, use the two-argument form of the package keyword
and bump the minimum version of Perl to 5.12 accordingly.

Clean up some other Module::Build details: get the synopsis from
the module POD, stop creating Makefile.PL files, and fix quoting
of the Pod::Simple dependency.  Also update the license from the
earlier change.