Fix internal section links and error handling
Links to other sections in the same document broke with the
refactoring of table of contents and navbar handling.  Fix that by
resolving links when possible and otherwise deferring them until
the end of the document, and then going back to add them in or
make them go away if the section was never found.

Output the document before raising an exception when POD errors
were seen.

Rename internal methods to start with a leading underscore (so the
only public method provided is new).