Treat an undef title as if it wasn't given
When title was passed as a constructor argument, it always
disabled NAME parsing, even if the title passed was undef.  This
makes constructing the arguments annoying, though, so treat undef
as if no title was specified.  To get the previous behavior, pass
an empty string as the value of the title option.